Add all existing classes into a package called duke.
