+++ src/main/java/duke/Duke.java
@@ -1,3 +1,4 @@
+package duke;
 import java.util.Scanner;
 import java.io.IOException;
 import java.io.File;
