+++ src/main/java/duke/Task.java
@@ -1,4 +1,4 @@
-
+package duke;
 /**
  * Encapsulates a description.
  */
