+++ src/main/java/duke/Ui.java
@@ -1,3 +1,4 @@
+package duke;
 /**
  * Handles interactions with the user.
  */
